@@ -1,18 +1,18 @@
 import 'dart:async';
 
 /// A variation of the Maybe monad with eager execution.
-abstract class Maybe<T> {
+abstract class Maybe<T extends Object> {
   /// Creates an instance of the monadic value.
-  factory Maybe(T value) => value == null ? Nothing<T>() : Just(value);
-
-  /// Maps the value to P. The [mapper] function must not return null.
-  Maybe<P> map<P>(P Function(T _) mapper);
+  factory Maybe(T? value) => value != null ? Just(value) : Nothing<T>();
 
   /// Maps the value to P.
-  Maybe<P> flatMap<P>(Maybe<P> Function(T _) mapper);
+  Maybe<P> map<P extends Object>(P Function(T value) mapper);
+
+  /// Maps the value to P.
+  Maybe<P> flatMap<P extends Object>(Maybe<P> Function(T value) mapper);
 
   /// Filter the value using the [predicate].
-  Maybe<T> where(bool Function(T _) predicate);
+  Maybe<T> where(bool Function(T value) predicate);
 
   /// Returns the wrapped value (if present), or the [defaultValue] (otherwise).
   T or(T defaultValue);
@@ -29,41 +29,42 @@ abstract class Maybe<T> {
   /// Returns the wrapped value (if present), or throws the result of the [producer] (otherwise).
   T orThrow(Object Function() producer);
 
-  /// Calls the [consumer] if the wrapped value is present. If the value is not present,
-  /// and the [otherwise] function is passed, calls it.
-  Maybe<T> ifPresent(void Function(T _) consumer, {void Function() otherwise});
+  /// Calls the [consumer] if the wrapped value is present.
+  Maybe<T> ifPresent(void Function(T value) consumer);
 
   /// Calls the [callback] function if the wrapped value is not present.
   Maybe<T> ifNothing(void Function() callback);
 
   /// Narrows the type to P if the value is present and has actually the type of P.
-  Maybe<P> type<P>();
+  Maybe<P> type<P extends Object>();
 
-  /// Merges the [other] value using the [merger] function.
-  Maybe<V> merge<V, R>(Maybe<R> other, V Function(T a, R b) merger);
+  /// If this and the other are both [Just] values, merges them using the [merger] function and returns [Just]<V>.
+  /// Otherwise returns [Nothing]<V>
+  Maybe<V> merge<R extends Object, V extends Object>(
+      Maybe<R> other, Merger<V, T, R> merger);
 
-  /// Returns the result of [next] if this is empty.
+  /// If this is [Nothing], returns the result of [next]. Otherwise return this..
   Maybe<T> fallback(Maybe<T> Function() next);
 
-  /// Returns the [next] Maybe if this is empty.
+  /// If this is [Nothing], returns [next]. Otherwise return this..
   Maybe<T> chain(Maybe<T> next);
 }
 
 /// Represents an existing non-null value of type T.
-class Just<T> implements Maybe<T> {
+class Just<T extends Object> implements Maybe<T> {
   /// Throws an [ArgumentError] if the value is null.
-  Just(this.value) {
-    ArgumentError.checkNotNull(value);
-  }
+  Just(this.value);
 
   /// The wrapped value. It is guaranteed to be non-null.
   final T value;
 
   @override
-  Maybe<P> map<P>(P Function(T _) mapper) => Maybe(mapper(value));
+  Maybe<P> map<P extends Object>(P Function(T value) mapper) =>
+      Maybe(mapper(value));
 
   @override
-  Maybe<P> flatMap<P>(Maybe<P> Function(T _) mapper) => mapper(value);
+  Maybe<P> flatMap<P extends Object>(Maybe<P> Function(T value) mapper) =>
+      mapper(value);
 
   @override
   T or(T defaultValue) => value;
@@ -81,7 +82,7 @@ class Just<T> implements Maybe<T> {
   T orThrow(Object Function() producer) => value;
 
   @override
-  Just<T> ifPresent(void Function(T _) consumer, {void Function() otherwise}) {
+  Just<T> ifPresent(void Function(T value) consumer) {
     consumer(value);
     return this;
   }
@@ -90,14 +91,16 @@ class Just<T> implements Maybe<T> {
   Just<T> ifNothing(void Function() callback) => this;
 
   @override
-  Maybe<T> where(bool Function(T _) predicate) =>
+  Maybe<T> where(bool Function(T value) predicate) =>
       predicate(value) ? this : Nothing<T>();
 
   @override
-  Maybe<P> type<P>() => value is P ? Just(value as P) : Nothing<P>();
+  Maybe<P> type<P extends Object>() =>
+      value is P ? Just(value as P) : Nothing<P>();
 
   @override
-  Maybe<V> merge<V, R>(Maybe<R> other, V Function(T a, R b) merger) =>
+  Maybe<V> merge<R extends Object, V extends Object>(
+          Maybe<R> other, Merger<V, T, R> merger) =>
       flatMap((a) => other.map((b) => merger(a, b)));
 
   @override
@@ -113,15 +116,19 @@ class Just<T> implements Maybe<T> {
   int get hashCode => value.hashCode;
 }
 
+typedef Merger<V extends Object, T extends Object, R extends Object> = V
+    Function(T a, R b);
+
 /// Represents a non-existing value of type T.
-class Nothing<T> implements Maybe<T> {
+class Nothing<T extends Object> implements Maybe<T> {
   const Nothing();
 
   @override
-  Nothing<P> map<P>(P Function(T _) mapper) => Nothing<P>();
+  Nothing<P> map<P extends Object>(P Function(T value) mapper) => Nothing<P>();
 
   @override
-  Nothing<P> flatMap<P>(Maybe<P> Function(T _) mapper) => Nothing<P>();
+  Nothing<P> flatMap<P extends Object>(Maybe<P> Function(T value) mapper) =>
+      Nothing<P>();
 
   @override
   T or(T defaultValue) => defaultValue;
@@ -139,11 +146,7 @@ class Nothing<T> implements Maybe<T> {
   T orThrow(Object Function() producer) => throw producer();
 
   @override
-  Nothing<T> ifPresent(void Function(T _) consumer,
-      {void Function() otherwise}) {
-    otherwise?.call();
-    return this;
-  }
+  Nothing<T> ifPresent(void Function(T value) consumer) => this;
 
   @override
   Nothing<T> ifNothing(void Function() callback) {
@@ -152,13 +155,14 @@ class Nothing<T> implements Maybe<T> {
   }
 
   @override
-  Nothing<T> where(bool Function(T _) predicate) => this;
+  Nothing<T> where(bool Function(T value) predicate) => this;
 
   @override
-  Nothing<P> type<P>() => Nothing<P>();
+  Nothing<P> type<P extends Object>() => Nothing<P>();
 
   @override
-  Nothing<V> merge<V, R>(Maybe<R> other, V Function(T a, R b) merger) =>
+  Nothing<V> merge<R extends Object, V extends Object>(
+          Maybe<R> other, Merger<V, T, R> merger) =>
       Nothing<V>();
 
   @override
